fix(webhooks): Refactor centralize webhooks code
Fixes: #2454
