+++ cl/recap/tasks.py
@@ -41,6 +41,7 @@ from rest_framework.status import (
 
 from cl.alerts.tasks import enqueue_docket_alert, send_alert_and_webhook
 from cl.api.models import WebhookEventType
+from cl.api.webhooks import send_recap_fetch_webhook_event
 from cl.celery_init import app
 from cl.corpus_importer.tasks import (
     download_pacer_pdf_by_rd,
@@ -88,7 +89,6 @@ from cl.recap.models import (
     PacerHtmlFiles,
     ProcessingQueue,
 )
-from cl.recap.utils import send_recap_fetch_webhook_event
 from cl.scrapers.tasks import extract_recap_pdf, extract_recap_pdf_base
 from cl.search.models import Court, Docket, DocketEntry, RECAPDocument
 from cl.search.tasks import add_items_to_solr, add_or_update_recap_docket
